removed unnecessary codes + modular code[#6]
