--- electron_GUI/addWindow.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Add Shopping List Item</title>
-</head> 
-<body>
-    <form>
-        <div>
-            <label>Enter Item</label>
-            <input type="text" id="item" autofocus>
-        </div>
-        <button type="submit">add Item</button>
-    </form>
-
-    <script>
-        const electron = require('electron');
-        const {ipcRenderer} = electron;
-
-        const form = document.querySelector('form');
-        form.addEventListener('submit', (e) => {
-            console.log('x');
-            e.preventDefault();  // stop default submit
-            const item = document.querySelector('#item').value;
-            ipcRenderer.send('item:add', item);  // like socketIO
-        });
-    </script>
-</body>
-</html>
